wip port add instance modal to client page
