+++ src/pages/client/components/ClientSideBar/ClientSideBar.tsx
@@ -1,5 +1,7 @@
 import './style.scss';
 
+import { useNavigate } from 'react-router-dom';
+
 import { useI18nContext } from '../../../../i18n/i18n-react';
 import SvgDefguadNavLogoCollapsed from '../../../../shared/components/svg/DefguardLogoCollapsed';
 import SvgDefguardLogoIcon from '../../../../shared/components/svg/DefguardLogoIcon';
@@ -7,8 +9,8 @@ import SvgDefguardLogoText from '../../../../shared/components/svg/DefguardLogoT
 import SvgIconNavConnections from '../../../../shared/components/svg/IconNavConnections';
 import { IconContainer } from '../../../../shared/defguard-ui/components/Layout/IconContainer/IconContainer';
 import SvgIconPlus from '../../../../shared/defguard-ui/components/svg/IconPlus';
+import { routes } from '../../../../shared/routes';
 import { useClientStore } from '../../hooks/useClientStore';
-import { useAddInstanceModal } from '../modals/AddInstanceModal/hooks/useAddInstanceModal';
 import { ClientBarItem } from './components/ClientBarItem/ClientBarItem';
 
 export const ClientSideBar = () => {
@@ -40,12 +42,14 @@ export const ClientSideBar = () => {
 
 const AddInstance = () => {
   const { LL } = useI18nContext();
-  const openAddInstanceModal = useAddInstanceModal((state) => state.open);
+  const navigate = useNavigate();
   return (
     <div
       id="add-instance"
       className="client-bar-item clickable"
-      onClick={() => openAddInstanceModal()}
+      onClick={() => {
+        navigate(routes.client.addInstance, { replace: true });
+      }}
     >
       <IconContainer>
         <SvgIconPlus />
