--- src/pages/client/components/LocationsList/components/LocationCardNoData/LocationCardNoData.tsx
@@ -1,13 +1,0 @@
-import './style.scss';
-
-import { useI18nContext } from '../../../../../../i18n/i18n-react';
-
-export const LocationCardNoData = () => {
-  const { LL } = useI18nContext();
-  return (
-    <div className="location-no-data">
-      <p>{LL.pages.client.locationNoData.title()}</p>
-      <p>{LL.pages.client.locationNoData.content()}</p>
-    </div>
-  );
-};
