+++ src/pages/client/pages/ClientAddInstancePage/components/AddInstanceFormCard/types.ts
@@ -1,0 +1,4 @@
+export type AddInstnaceInitResponse = {
+  url: string;
+  cookie: string;
+};
