+++ src/pages/client/pages/ClientInstancePage/components/LocationUsageChart/types.ts
@@ -1,0 +1,4 @@
+export enum LocationUsageChartType {
+  BAR,
+  LINE,
+}
