+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationCardConnectButton/style.scss
