+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationCardInfo/style.scss
