+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationCardNoData/style.scss
