+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationCardTitle/style.scss
