+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationsDetailView/style.scss
