+++ src/pages/client/pages/ClientInstancePage/components/LocationsList/components/LocationsGridView/style.scss
