+++ src/pages/client/style.scss
@@ -1,6 +1,6 @@
 @use '@scssutils' as *;
 
-#client-page {
+.client-page {
   width: 100%;
   box-sizing: border-box;
   min-height: 100vh;
@@ -11,38 +11,6 @@
 
   @include media-breakpoint-up(lg) {
     padding: 64px 75px 64px calc(75px + 270px);
-  }
-
-  & > header {
-    width: 100%;
-    display: flex;
-    flex-flow: row wrap;
-    align-items: center;
-    column-gap: 30px;
-    row-gap: 25px;
-    padding-bottom: 30px;
-
-    @include media-breakpoint-up(lg) {
-      padding-bottom: 60px;
-    }
-
-    & > h1 {
-      @include typography(app-title);
-      color: var(--text-body-primary);
-    }
-
-    .options {
-      display: flex;
-      flex-flow: row nowrap;
-      align-self: center;
-      justify-content: flex-start;
-      column-gap: 20px;
-    }
-
-    .select-container {
-      height: 40px;
-      width: 134px;
-    }
   }
 
   a,
